refactor: extract game engine from Game.js into gameEngine.js
Move all pure game logic (combat, abilities, DoT, status effects,
initiative, victory check, AI decision, deck creation) into a new
gameEngine.js module with zero React dependencies.

Game.js (569 lines) is now a thin React wrapper handling state,
animations, and UI rendering. gameEngine.js (419 lines) contains
all game rules and can run independently — enabling future server-side
use (Cloudflare Workers), deterministic testing, and story mode reuse.

All 49 tests pass with updated imports.
